metrics/requirements-dev.txt
ruff==0.6.4
pytest==8.3.3
Faker==28.4.1
eth-typing==4.1.0
pytest-aiohttp==1.0.5